build image on TC
