+++ Dockerfile
@@ -1,8 +1,7 @@
 FROM tiangolo/uwsgi-nginx-flask:python3.6
 
-COPY ./requirements.txt .
+COPY ./requirements-docker.txt requirements.txt
 RUN pip3 install -r requirements.txt
 
 ENV STATIC_INDEX 1
 COPY ./src /app
-COPY config.final.json /app
