+++ requirements-docker.txt
@@ -1,0 +1,2 @@
+flask>=0.12.4
+requests
