+++ requirements.txt
@@ -1,3 +1,1 @@
-flask>=0.12.4
-requests
 hvac
