+++ tests/test_branch_matching.py
@@ -1,22 +1,22 @@
-from src.app.main import get_issue_id
+from src.app.YouTrackHelper import YouTrackHelper
 
 
 def test_can_match_old_branch_names():
-    result = get_issue_id("i637")
+    result = YouTrackHelper.get_issue_id("i637")
     assert result == "VIMC-637"
 
-    result = get_issue_id("i123_something")
+    result = YouTrackHelper.get_issue_id("i123_something")
     assert result == "VIMC-123"
 
 
 def test_can_match_new_branch_names():
-    result = get_issue_id("VIMC-637")
+    result = YouTrackHelper.get_issue_id("VIMC-637")
     assert result == "VIMC-637"
 
-    result = get_issue_id("mrc-123_something")
+    result = YouTrackHelper.get_issue_id("mrc-123_something")
     assert result == "mrc-123"
 
 
 def test_does_not_match_arbitrary_branch_names():
-    result = get_issue_id("refactor")
+    result = YouTrackHelper.get_issue_id("refactor")
     assert result is None
